requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
sqlalchemy>=1.4.23
alembic>=1.7.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
redis>=4.0.2
prometheus-client>=0.11.0
dependency-injector>=4.39.1
python-json-logger>=2.0.2
PyYAML>=5.4.1
aiohttp>=3.8.1
numpy>=1.21.2
torch>=1.9.0
faiss-cpu>=1.7.2
transformers>=4.9.2
python-multipart>=0.0.5
psutil>=5.8.0
GPUtil>=1.4.0